refactor: unify history ui manager names
