+++ src/historyView/index.html
@@ -19,10 +19,10 @@
       {
         "imports": {
           "./modules/domHandlers.js": "${domHandlersUri}",
-          "./modules/uiManagers/historyEvents.js": "${historyEventsUri}",
-          "./modules/uiManagers/historyRenderer.js": "${historyRendererUri}",
+          "./modules/uiManagers/HistoryEvents.js": "${historyEventsUri}",
+          "./modules/uiManagers/HistoryRenderer.js": "${historyRendererUri}",
           "./modules/historyViewState.js": "${historyViewStateUri}",
-          "./modules/uiManagers/searchManager.js": "${searchManagerUri}",
+          "./modules/uiManagers/SearchManager.js": "${searchManagerUri}",
           "./modules/messageHandlers.js": "${messageHandlersUri}",
           "@common/webviewState.js": "${webviewStateUri}",
           "@common/webview/commands.js": "${commandsUri}",
