+++ src/historyView/modules/uiManagers/HistoryEvents.js
