+++ src/historyView/modules/uiManagers/SearchManager.js
